Rename clients to gateways (#101)
