+++ gateway/controller_test.go
@@ -1,4 +1,4 @@
-package clients
+package gateway
 
 import (
 	"encoding/json"
